feat: 툴팁 작업 완료, 워닝 제거 방안 찾는 중
