+++ screen/Home.js
@@ -14,14 +14,12 @@ import { TouchableOpacity, View } from 'react-native';
 import { useRecoilValue, useRecoilState } from 'recoil';
 import userInfo from '../recoil/userInfo';
 import DrawerState from '../recoil/drawer';
-import TooltipState from '../recoil/tooltip';
-import BasicToolTip from '../components/BasicToolTip';
+import HeaderToolTip from '../components/HeaderToolTip';
 
 const Home = () => {
   const Tab = createBottomTabNavigator();
   const { userId } = useRecoilValue(userInfo);
   const [isOpen, setIsOpen] = useRecoilState(DrawerState);
-  const [toolTip, setToolTip] = useRecoilState(TooltipState);
 
   return (
     <NavigationContainer>
@@ -84,25 +82,7 @@ const Home = () => {
             ),
             headerRight: () => {
               if (route.name !== 'ListHome') {
-                return (
-                  <TouchableOpacity
-                    style={{ marginRight: 16 }}
-                    onPress={() => {
-                      const props = {
-                        show: !toolTip.show,
-                        routeType: route.name,
-                      };
-                      //route.name으로 상태 구분하기
-                      setToolTip(props);
-                    }}
-                  >
-                    <MaterialCommunityIcons
-                      name="information-outline"
-                      size={24}
-                      color={theme.white}
-                    />
-                  </TouchableOpacity>
-                );
+                return <HeaderToolTip routeName={route.name} />;
               }
             },
           })}
@@ -163,7 +143,6 @@ const Home = () => {
             }}
           />
         </Tab.Navigator>
-        <BasicToolTip />
         <GlobalModal />
       </Drawer>
     </NavigationContainer>
